+++ packages/theme/components/Blog/BlogInfo.vue
@@ -19,7 +19,11 @@
             :key="article.path"
             :delay="(index + 1) * 0.08"
           >
-            <li class="sticky-article" @click="$router.push(article.path)" v-text="article.title" />
+            <li
+              class="sticky-article"
+              @click="$router.push(article.path)"
+              v-text="article.title"
+            />
           </MyTransition>
         </ul>
       </div>
@@ -28,7 +32,11 @@
     <!-- 分类 -->
     <MyTransition :delay="0.12">
       <div class="category-wrapper">
-        <div v-if="$category.list.length !== 0" class="title" @click="$router.push('/category/')">
+        <div
+          v-if="$category.list.length !== 0"
+          class="title"
+          @click="$router.push('/category/')"
+        >
           <CategoryIcon />
           <span class="num">{{ $category.list.length }}</span>
           {{ i18n.category }}
@@ -43,7 +51,11 @@
     <!-- 标签 -->
     <MyTransition :delay="0.2">
       <div class="tag-wrapper">
-        <div v-if="$tag.list.length !== 0" class="title" @click="$router.push('/tag/')">
+        <div
+          v-if="$tag.list.length !== 0"
+          class="title"
+          @click="$router.push('/tag/')"
+        >
           <TagIcon />
           <span class="num">{{ $tag.list.length }}</span>
           {{ i18n.tag }}