+++ packages/theme/layouts/Blog.vue
@@ -4,7 +4,10 @@
       <BlogInfo />
     </template>
 
-    <Password v-if="currentPathEncrypted && !globalEncrypted" @password-verify="setPassword" />
+    <Password
+      v-if="currentPathEncrypted && !globalEncrypted"
+      @password-verify="setPassword"
+    />
     <BlogPage v-else />
   </Common>
 </template>